finished moving everything to separate files
