+++ lua/plugins/autocomplete/init.lua
@@ -1,9 +1,4 @@
 return {
-    { -- Snippets
-        'L3MON4D3/LuaSnip',
-        lazy = true,
-        build = vim.g.isnix and "nix-shell -p gnumake --run 'make install_jsregexp'" or "make install_jsregexp",
-        dependencies = 'rafamadriz/friendly-snippets',
-        opts = { history = true, updateevents = 'TextChanged,TextChangedI' },
-    },
+    { 'tpope/vim-surround', keys = { 'cs', 'ds', 'ys', 'cst', 'dsf', 'dst' }, dependencies = 'tpope/vim-repeat' },
+
 }
