+++ lua/plugins/ui/init.lua
@@ -1,4 +1,7 @@
 return {
     { 'j-hui/fidget.nvim', event = "VeryLazy", opts = {} },
     { 'kevinhwang91/nvim-ufo', dependencies = 'kevinhwang91/promise-async', opts = {} },
+    { 'folke/which-key.nvim', event = 'VeryLazy' },
+    { "karb94/neoscroll.nvim", opts = {} },
+
 }
